Renaming profile component
